Removed FilterableAggregation class
