+++ pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/function/AggregationFunction.java
@@ -129,6 +129,12 @@ public interface AggregationFunction<IntermediateResult, FinalResult extends Com
    */
   FinalResult extractFinalResult(IntermediateResult intermediateResult);
 
+  /** Set if this aggregation is filtered */
+  void setFilteredAggregation(boolean isFilteredAggregation);
+
+  /** @return If the aggregation is filtered */
+  boolean isFilteredAggregation();
+
   /** @return Description of this operator for Explain Plan */
   String toExplainString();
 }
